modernize implementation and tests; leave current API as-is
This is a bunch of cleanups to get the implementation and tests
up to our modern standards as much as possible.  The original
source (from kelredd-pruview) was written circa 2007 and before
Assert was even a thing.  The modernizations includes things
like:

* organization and requiring of files
* style (as much as possible without a wholesale rewrite)
* use assert for tests
* rename `ffyml` bin to `pruview-ffyml`
* lookup bin path using rubygems api
* drop support for converting to FLV (and the flvtool2 dep) - FLV
  is an outdated video format

This is all to get the implementation in a better place for
ongoing maintenance and support.